# Conflicts:
#	app/src/main/java/saneforce/sanzen/activity/call/adapter/detailing/PlaySlideDetailedAdapter.java
#	app/src/main/java/saneforce/sanzen/storage/SharedPref.java

+++ app/src/main/java/saneforce/sanzen/activity/call/fragments/signature/SignatureFragment1.java
@@ -72,7 +72,6 @@ public class SignatureFragment1 extends Fragment {
         callSignCaptureImage = new ArrayList<>();
     }
 
-
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
@@ -155,10 +154,9 @@ public class SignatureFragment1 extends Fragment {
         }
     }
 
-
     public void clearSignature() {
         if (signatureCanvas != null) {
-            if (!callSignCaptureImage.isEmpty()) {
+            if(callSignCaptureImage != null && !callSignCaptureImage.isEmpty()) {
                 File fileDelete = new File(callSignCaptureImage.get(0).getFilepath());
                 if (fileDelete.exists()) {
                     if (fileDelete.delete()) {
@@ -166,6 +164,8 @@ public class SignatureFragment1 extends Fragment {
                         callSignCaptureImage.clear();
                     }
                 }
+            } else {
+                callSignCaptureImage = new ArrayList<>();
             }
             signatureCanvas.clearCanvas();
             signatureCanvas.setSignaturePath(new Path());
@@ -286,7 +286,6 @@ public class SignatureFragment1 extends Fragment {
         }
     }
 
-
     public void loadImageFromLocal(String fileName) {
         if (/*!filePath.isEmpty() && */!imageName.isEmpty()) {
             File file = new File(context.getExternalFilesDir(null) + "/Signature/", fileName);
@@ -301,8 +300,6 @@ public class SignatureFragment1 extends Fragment {
             Log.d("TAG", "instance initializer: file path is empty");
         }
     }
-
-
 /*    public void loadImageFromLocal() {
         if (!filePath.isEmpty()) {
             File file = new File(filePath);
